% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_riddle.R
\name{error_response}
\alias{error_response}
\title{Response for a solutions that produce errors}
\usage{
error_response(x, solution_output, input_env, test_i, test_n)
}
\description{
Response for a solutions that produce errors
}
