% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_riddle.R
\name{eval_with_timeout}
\alias{eval_with_timeout}
\title{Evaluate and memoise solution execution}
\usage{
eval_with_timeout(x, timeout = Inf, envir = parent.frame())
}
\description{
Evaluate and memoise solution execution
}
