% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_riddle.R
\name{grace_timeouts_response}
\alias{grace_timeouts_response}
\title{Response for a solution that times out only a grace period}
\usage{
grace_timeouts_response(
  user_output,
  solution_output,
  input_env,
  test_i,
  test_n,
  time_frac
)
}
\description{
Response for a solution that times out only a grace period
}
