% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_riddle.R
\name{grade_riddle}
\alias{grade_riddle}
\title{function for generating an exercise checker for a puzzlr challenge}
\usage{
grade_riddle(
  user_code,
  solution,
  test_inputs,
  test_timeouts = Inf,
  grace_timeouts = test_timeouts * 5,
  test_details = rep(list(NULL), length(test_inputs)),
  quoted = FALSE
)
}
\arguments{
\item{solution}{solution code}

\item{envs}{environments in which solution should be evaluated. these should
contain the necessary input variables for the challenge.}
}
\value{
A shiny ui element displaying output from grading the question
  solution.
}
\description{
function for generating an exercise checker for a puzzlr challenge
}
