% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_riddle.R
\name{incorrect_solution_response}
\alias{incorrect_solution_response}
\title{Response for an incorrect solution}
\usage{
incorrect_solution_response(
  user_output,
  solution_output,
  input_env,
  test_i,
  test_n,
  err
)
}
\description{
Response for an incorrect solution
}
