% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_timeout_error}
\alias{is_timeout_error}
\title{Compare an error against a timeout error from setTimeLimit}
\usage{
is_timeout_error(e)
}
\arguments{
\item{e}{an error to compare}
}
\value{
a logical indicating whether e is a setTimeLimit timeout error
}
\description{
Compare an error against a timeout error from setTimeLimit
}
