% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\docType{data}
\name{opts}
\alias{opts}
\title{Getters and setters for riddlr options}
\format{
An object of class \code{list} of length 2.
}
\usage{
opts
}
\description{
\describe{
  \item{onSubmit}{
    Expects a function expecting arguments \code{code}, \code{duration} and
    \code{response}. This function is called after a question solution is
    submitted and graded and can be used as a callback for additional
    functionality or logging.
  }
}
}
\examples{
opts$set(onSubmit = function(code, duration, response) message("submitted"))

}
\keyword{datasets}
