% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_riddle.R
\name{parse_error_response}
\alias{parse_error_response}
\title{Response for a solutions that produce errors}
\usage{
parse_error_response(x, ...)
}
\description{
Response for a solutions that produce errors
}
