% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riddlr_env.R
\docType{data}
\name{riddlr_base_env}
\alias{riddlr_base_env}
\title{Build a base environment for solutions to be evaluated within}
\format{An object of class \code{environment} of length 0.}
\usage{
riddlr_base_env
}
\description{
Build a base environment for solutions to be evaluated within
}
\keyword{datasets}
