% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{timeout_error_cpu}
\alias{timeout_error_cpu}
\title{compute a timeout error from setTimeLimit}
\format{An object of class \code{NULL} of length 0.}
\usage{
timeout_error_cpu
}
\description{
compute a timeout error from setTimeLimit
}
\keyword{datasets}
