% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_riddle.R
\name{timeout_response}
\alias{timeout_response}
\title{Response for a solution that times out}
\usage{
timeout_response(x, solution_output, input_env, test_i, test_n)
}
\description{
Response for a solution that times out
}
