% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_riddle.R
\name{update_test_progress_html}
\alias{update_test_progress_html}
\title{Format test results as html for updating progress text}
\usage{
update_test_progress_html(test_results, reactive_output)
}
\description{
Format test results as html for updating progress text
}
