% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riddlr_env.R
\name{with_env_cleanup}
\alias{with_env_cleanup}
\title{Helper to ensure that any library calls get undone by the end of evaluation}
\usage{
with_env_cleanup(expr, envir = parent.frame())
}
\description{
Helper to ensure that any library calls get undone by the end of evaluation
}
